requirements.txt
wheel 
djangorestframework
django-filter
drf-yasg
djangorestframework-simplejwt
Pillow

psycopg2-binary
python-decouple
python-slugify

django-phonenumber-field